refactor: move credential utils in dedicated file
